requirements.txt
appdirs==1.4.3
future==0.16.0
IPy==0.83
Logbook==1.0.0
netifaces==0.10.5
olefile==0.44
packaging==16.8
Pillow==4.1.0
PyInstaller==3.2.1
pyparsing==2.2.0
pyperclip==1.5.27
pypiwin32==219
PyYAML==3.12
qrcode==5.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
tornado==4.5
